pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "carbonlens"
version = "0.2.0"
description = "Production-ready emissions analytics platform combining EDGAR v2024 datasets, Model Context Protocol (MCP) integration, and LLM-powered natural language queries. Analyze multi-sector greenhouse gas data with geographic granularity from cities to countries, 2000-2024."
readme = "README.md"
requires-python = ">=3.11,<3.12"

# Pin versions for reproducible builds
dependencies = [
  "altair>=5.5.0",
  "duckdb>=1.4.1",
  "fastapi==0.117.1",
  "pydantic==2.11.9",
  "geopandas==1.1.1",
  "h5netcdf==1.6.4",
  "h5py==3.14.0",
  "netcdf4>=1.6.0", # Added for NetCDF file processing
  "numpy==2.3.3",
  "openai>=2.2.0",
  "pandas==2.3.2",
  "plotly==6.3.0",
  "pyarrow==21.0.0",
  "pyproj==3.7.2",
  "pyogrio==0.11.1",
  "python-dotenv==1.1.1",
  "requests==2.32.5",
  "httpx==0.28.1",
  "tenacity==9.1.2",
  "tqdm==4.67.1",
  "click==8.3.0",
  "rtree==1.4.1",
  "shapely==2.1.1",
  "pydeck==0.9.1",
  "streamlit==1.50.0",
  "uvicorn[standard]==0.37.0",
  "watchfiles==1.1.0",
  "xarray==2025.9.0",
  "mcp>=1.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "ruff==0.6.9",
  "black==24.8.0",
]

# Optional extras for database checking/reporting utilities
checker = [
  "sqlalchemy>=1.4.0",
  "matplotlib>=3.5.0",
  "seaborn>=0.11.0",
  "jinja2>=3.0.0",
  "pytest-cov>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
